# Core: Skip event sweep for players which did not receive any advancement in the previous iteration