--- a/BaseClasses.py
+++ b/BaseClasses.py
@@ -711,7 +711,7 @@ class CollectionState():
 
         # The first iteration must check the locations for all players because it is not known which players might have
         # reachable locations.
-        players_to_check = set(self.multiworld.player_ids)
+        players_to_check = set(self.multiworld.regions.location_cache.keys())
         while players_to_check:
             # Iterable of all locations belonging to the players to check.
             loc_gen: Iterable[Location] = itertools.chain.from_iterable(