… docs/world api.md
… worlds/AutoWorld.py