## What is this fixing or adding?
Skip checking location accessibility for locations belonging to a player which did not receive any advancement (`collection_state.collect(item) == True` where `item.player == player`) in the last iteration. This can significantly decrease generation time by not having to check accessibility of some players' locations each iteration.

The majority of generation time is usually taken up by checking location accessibility, so reducing the number of locations that need to be checked can have a great effect on generation time. The percentage reduction in generation time with this patch appears to be greater the larger the multiworld. Generating with `python -O .\Generate.py --skip_output --seed 1` with 4 copies of each template yaml, except A Link to the Past, SMZ3, Pokemon R/B and Yu-gi-oh!, resulted in generation time being reduced by 57% (from ~1193s to ~505s, though 5 generations were run simultaneously to more quickly produce results, so the actual generation times are expected to be lower) [TODO: Re-time with individual generations instead of 5 simultaneously]

A potential caveat with this change is that it assumes that all logic for a player's world only depends on the items belonging to that player. It is not clear to me if this is a strict requirement for Archipelago worlds to have, however, all supported world appear to meet the requirement.

## How was this tested?
Fixed seeds were generated, with and without this patch, with different sets of template yamls in the Players directory, using games known to have deterministic generation. The output from each set of template yamls was seen to match.

Seeds using item links have also been tested.

---

Generated with `python -O .\Generate.py --skip_output --seed 1`
[TODO: Time some more generations of varying sizes and put the results into the table]
| Yamls | Before Patch/s | After Patch/s |
| -- | -- | -- |
| 4 of each template, except A Link to the Past, SMZ3, Pokemon R/B and Yu-gi-oh! | 1193 | 505 |

---

The `event_filter` functions may be incorrect, see #3680.

This is a draft until I have more generation timings to compare and do more self-review.