@Mysteryem 2024-07-25
> i think my biggest comment here is that i'm a bit confused why this is using sets. i don't see anything in particular that seems like it needs to be a set (instead of a list). sets are more expensive to create and slower to iterate over. the only bit in the old code as far as i can see that justifies using a set is `locations -= reachable_events`. it also used to not be guaranteed for locations to have unique names, but that's enforced nowadays so we don't have to worry about duplicates

It should be possible to have a list of locations, where inaccessible locations are iterated into a new list that replaces the old list after the iteration. Another alternative could be to iterate the list in reverse along with a reversed enumeration, that way, reachable locations could be popped by index. I'll investigate some list-based options.