@Mysteryem 2024-07-27
I think there's some more places where performance can be improved, but some good performance improvements have been made since I learned more about the cache for region accessibility, and I have updated the table. This has also increased the scope of the PR to a more general performance improvement, hopefully it will not increase much further because I should probably start thinking about splitting up the PR if it does, since I don't want to make it too difficult for whoever has to review this once its ready.

Given the growing size of the `sweep_for_events()` function, it might be good idea to split it up parts of it into smaller, private functions, perhaps one for filtering locations, another for computing `logic_dependent_players` and the last for performing the sweep loop.

Replacing the sets with lists was a good call and reduced generation time by about 4% percent.