@Mysteryem 2024-07-31
It's possible to squeeze in maybe a further 1% reduction in duration by pre-binding all instance methods and pre-getting all instance attributes used in loops and storing them as local variables, but this makes the function longer and less readable, so I have not done so.
e.g. only get `self.events` once and store it as a local variable and only get `accessible_locations.append` once and store it as a local variable etc.