@Mysteryem 2024-08-05
I've noticed that `MultiWorld.can_beat_game()` runs pretty much an identical sweep to `CollectionState.sweep_for_events()`, except it returns early once every player's game is beatable, so I think it can benefit from the same performance changes made here. I wonder if I can make the code generic enough to be used by both functions without a noticeable loss of performance.