@Mysteryem 2024-08-12
The decrease in generation duration for the 2-of-each-yaml case seems a bit high, so I want to try comparing some other seeds, in-case there was something about that seed that the original code particularly struggled with. Maybe I'll profile generating with 4 of each template yaml again too.

Aside from that, I think the implementation is all here now. If there's no fast movement on this PR, I'll probably check in in a week or two and do some more self-review once I've lost familiarity with some of the changes.